refactor: move type4 test to v1 compressor package
